bit of a mess, reverted to incluide PointDataset
