+++ autolens/point/dataset.py
@@ -1,7 +1,3 @@
-import json
-import numpy as np
-import os
-from os import path
 from typing import List, Tuple, Optional, Union
 
 import autoarray as aa
