+++ autolens/point/fit/positions/image/abstract.py
@@ -4,10 +4,8 @@ import autoarray as aa
 import autogalaxy as ag
 
 from autolens.point.solver import PointSolver
-from autolens.point.fit_point.positions.abstract import AbstractFitPositions
+from autolens.point.fit.positions.abstract import AbstractFitPositions
 from autolens.lens.tracer import Tracer
-
-from autolens import exc
 
 
 class AbstractFitPositionsImagePair(AbstractFitPositions):
