+++ autolens/point/fit/positions/source/max_separation.py
