+++ autolens/point/plot/fit_point_plotters.py
@@ -2,7 +2,7 @@ from autoarray.plot.abstract_plotters import AbstractPlotter
 
 import autogalaxy.plot as aplt
 
-from autolens.point.fit_point.point_dataset import FitPointDataset
+from autolens.point.fit.dataset import FitPointDataset
 
 
 class FitPointDatasetPlotter(AbstractPlotter):
