+++ test_autolens/point/fit/positions/image/test_abstract.py
