+++ test_autolens/point/fit/positions/image/test_pair.py
