+++ test_autolens/point/fit/positions/image/test_pair_all.py
